♻️ Refactor: major backend cleanup
