+++ backend/index.ts
@@ -1,48 +1,38 @@
-import express from "express";
-import cors from "cors";
-import { S3, ListObjectsV2Command } from "@aws-sdk/client-s3";
-import multer from "multer";
-import multerS3 from "multer-s3";
+import { Elysia, t } from "elysia";
+import { cors } from "@elysiajs/cors";
+import { S3, ListObjectsV2Command, PutObjectCommand } from "@aws-sdk/client-s3";
 import { analyzeImage } from "./services/claudeService";
 import { storeAnalysis } from "./services/dbService";
 import { generateSearchEmbedding } from "./services/openaiService";
 import searchPosts from "./services/dbService";
 
-interface SearchResult {
-  id: string;
-  analysis: string;
-  similarity: number;
-  created_at: Date;
-}
+const PORT = Number(process.env.PORT) || 3001;
+const BUCKET_NAME = process.env.BUCKET_NAME || "greengrep";
+const AWS_REGION = process.env.AWS_REGION || "us-east-1";
+const AWS_ACCESS_KEY_ID = process.env.AWS_ACCESS_KEY_ID || "";
+const AWS_SECRET_ACCESS_KEY = process.env.AWS_SECRET_ACCESS_KEY || "";
 
-const app = express();
-const PORT = process.env.PORT || 3001;
-
-// Apply CORS middleware
-app.use(cors());
-
-// Body parser middleware (JSON)
-app.use(express.json());
-
-// Initialize S3 client
 const s3 = new S3({
-  region: "us-east-1",
+  region: AWS_REGION,
   credentials: {
-    accessKeyId: process.env.AWS_ACCESS_KEY_ID || "",
-    secretAccessKey: process.env.AWS_SECRET_ACCESS_KEY || "",
+    accessKeyId: AWS_ACCESS_KEY_ID,
+    secretAccessKey: AWS_SECRET_ACCESS_KEY,
   },
-  // logger: console,
 });
 
-const BUCKET_NAME = "greengrep";
+const app = new Elysia()
+  .use(cors())
+  // Global error handler: log the error and respond with a generic message
+  .onError(({ error }) => {
+    console.error("Server error:", error);
+    return { error: "Internal Server Error" };
+  });
 
-// List all files in bucket
-// Example usage
-// curl -X GET http://localhost:3001/api/files
-app.get("/api/files", async (req, res) => {
+// List all files in S3 bucket
+app.get("/api/files", async ({ query, error }) => {
   try {
-    const limit = parseInt(req.query.limit as string) || 20;
-    const cursor = req.query.cursor as string;
+    const limit = parseInt(query.limit ?? "20", 10);
+    const cursor = query.cursor;
 
     const command = new ListObjectsV2Command({
       Bucket: BUCKET_NAME,
@@ -51,128 +41,110 @@ app.get("/api/files", async (req, res) => {
     });
 
     const response = await s3.send(command);
+    const files = (response.Contents || []).map(
+      ({ Key, LastModified, Size }) => ({
+        key: Key,
+        lastModified: LastModified,
+        size: Size,
+        url: `https://${BUCKET_NAME}.s3.amazonaws.com/${Key}`,
+      })
+    );
 
-    const files =
-      response.Contents?.map((object) => ({
-        key: object.Key,
-        lastModified: object.LastModified,
-        size: object.Size,
-        url: `https://${BUCKET_NAME}.s3.amazonaws.com/${object.Key}`,
-      })) || [];
-
-    res.json({
+    return {
       files,
       nextCursor: response.NextContinuationToken,
       hasMore: response.IsTruncated,
-    });
-  } catch (error) {
-    console.error("List files error:", error);
-    res.status(500).json({ error: "Failed to list files" });
+    };
+  } catch (err) {
+    console.error("List files error:", err);
+    return error(500, { error: "Failed to list files" });
   }
-});
-
-const upload = multer({
-  storage: multerS3({
-    s3: s3,
-    bucket: BUCKET_NAME,
-    metadata: (req, file, cb) => {
-      cb(null, { fieldName: file.fieldname });
-    },
-    key: (req, file, cb) => {
-      // Use the original file name or generate a unique one
-      const fileName = `${Date.now().toString()}-${file.originalname}`;
-      cb(null, fileName);
-    },
-  }),
-  fileFilter: (req, file, cb) => {
-    // Accept only jpg, jpeg, and png files
-    if (file.mimetype === "image/jpeg" || file.mimetype === "image/png") {
-      cb(null, true);
-    } else {
-      cb(new Error("Only JPG and PNG files are allowed"));
-    }
-  },
 });
 
 // Upload file endpoint
-// Example usage:
-// curl -X POST -F "file=@/path/to/your/file.jpg" http://localhost:3001/api/upload
-app.post("/api/upload", upload.single("file"), async (req, res) => {
-  try {
-    if (!req.file) {
-      throw new Error("No file uploaded");
+// Client should POST with multipart/form-data, field "file"
+app.post(
+  "/api/upload",
+  async ({ body, error }) => {
+    const { file } = body as { file: File };
+
+    if (!file) {
+      return error(400, { error: "No file uploaded" });
     }
 
-    const s3File = req.file as Express.MulterS3.File;
-    const fileUrl = s3File.location;
+    const uniqueFileName = `${Date.now()}-${file.name}`;
 
-    // Analyze the image using Claude
-    const analysis = await analyzeImage(
-      fileUrl,
-      s3File.originalname,
-      s3File.size.toString(),
-      s3File.mimetype
-    );
+    try {
+      // Convert File to Buffer
+      const arrayBuffer = await file.arrayBuffer();
+      const buffer = Buffer.from(arrayBuffer);
 
-    // Store the analysis in the database
-    const storedPosts = await storeAnalysis(analysis);
+      // Upload file to S3
+      await s3.send(
+        new PutObjectCommand({
+          Bucket: BUCKET_NAME,
+          Key: uniqueFileName,
+          Body: buffer,
+          ContentType: file.type || "application/octet-stream",
+        })
+      );
 
-    res.json({
-      message: "File uploaded and analyzed successfully",
-      fileUrl,
-      analysis,
-      storedPosts,
-    });
-  } catch (error: Error | unknown) {
-    const errorMessage =
-      error instanceof Error ? error.message : "Failed to upload file";
-    console.error("Upload error:", error);
-    res.status(500).json({
-      error: errorMessage,
-    });
+      const fileUrl = `https://${BUCKET_NAME}.s3.amazonaws.com/${uniqueFileName}`;
+
+      // Analyze the image using Claude
+      const analysis = await analyzeImage(
+        fileUrl,
+        file.name,
+        file.size.toString(),
+        file.type
+      );
+
+      // Store the analysis in the database
+      const storedPosts = await storeAnalysis(analysis);
+
+      return {
+        message: "File uploaded and analyzed successfully",
+        fileUrl,
+        analysis,
+        storedPosts,
+      };
+    } catch (err) {
+      console.error("Upload error:", err);
+      return error(500, { error: "Failed to upload file" });
+    }
+  },
+  {
+    // Validation schema
+    body: t.Object({
+      file: t.File({
+        type: ["image/jpeg", "image/png"], // Allowed MIME types
+        maxSize: "5m", // Max file size is 5 MB
+      }),
+    }),
   }
-});
+);
 
-// Add new endpoint for vector similarity search
-app.get("/api/vector-search", (async (
-  req: express.Request,
-  res: express.Response
-) => {
-  const { q: searchQuery } = req.query;
-
-  if (!searchQuery || typeof searchQuery !== "string") {
-    return res.status(400).json({
-      error: "Search query is required",
-    });
+// Vector similarity search
+app.get("/api/vector-search", async ({ query, error }) => {
+  const searchQuery = query.q;
+  if (!searchQuery) {
+    return error(400, { error: "Search query is required" });
   }
 
   try {
     const embedding = await generateSearchEmbedding(searchQuery);
     const results = await searchPosts(embedding);
 
-    res.json({
+    return {
       query: searchQuery,
       results,
-    });
-  } catch (error: Error | unknown) {
-    const errorMessage =
-      error instanceof Error ? error.message : "Search failed";
-    console.error("Vector search error:", error);
-    res.status(500).json({
-      error: errorMessage,
-    });
+    };
+  } catch (err) {
+    console.error("Vector search error:", err);
+    return error(500, { error: "Search failed" });
   }
-}) as express.RequestHandler<
-  Record<string, never>,
-  | { query: string; results: SearchResult[]; embedding: number[] }
-  | { error: string },
-  never,
-  { q: string }
->);
-
-// Start server
-app.listen(PORT, () => {
-  console.log(`Server is running on http://localhost:${PORT}`);
 });
 
-console.log("Hello from Bun!");
+app.listen(PORT, () => {
+  console.log(`🦊 Elysia is running on http://localhost:${PORT}`);
+});
