+++ backend/services/claudeService.ts
@@ -1,35 +1,10 @@
 import { Anthropic } from "@anthropic-ai/sdk";
+import type { ImageAnalysis } from "./types";
 
 // Initialize Claude client
 const anthropic = new Anthropic({
   apiKey: process.env.ANTHROPIC_API_KEY,
 });
-
-// Define types based on the example JSON structure
-interface ImagePost {
-  post_id: string;
-  board: string | null;
-  timestamp: string;
-  poster: string;
-  has_image: boolean;
-  is_nsfw: boolean;
-  url?: string;
-  embedded_image?: {
-    filename?: string;
-    size?: string;
-    format?: string;
-    dimensions?: string;
-    description?: string;
-  };
-  content: {
-    greentext: string[];
-    text: string[];
-  };
-}
-
-interface ImageAnalysis {
-  posts: ImagePost[];
-}
 
 export async function analyzeImage(
   imageUrl: string,
@@ -47,19 +22,19 @@ export async function analyzeImage(
     console.log("Format:", format);
     console.log("Base64 length:", base64Image.length);
 
-    // Convert format string to proper media type
-    const mediaType =
-      format.toLowerCase() === "jpeg" || format.toLowerCase() === "image/jpeg"
-        ? "image/jpeg"
-        : format.toLowerCase() === "png" || format.toLowerCase() === "image/png"
-          ? "image/png"
-          : format.toLowerCase() === "gif" ||
-              format.toLowerCase() === "image/gif"
-            ? "image/gif"
-            : format.toLowerCase() === "webp" ||
-                format.toLowerCase() === "image/webp"
-              ? "image/webp"
-              : "image/jpeg";
+    // Media type mapping
+    const mediaTypeMap: { [key: string]: string } = {
+      jpeg: "image/jpeg",
+      "image/jpeg": "image/jpeg",
+      png: "image/png",
+      "image/png": "image/png",
+      gif: "image/gif",
+      "image/gif": "image/gif",
+      webp: "image/webp",
+      "image/webp": "image/webp",
+    };
+
+    const mediaType = mediaTypeMap[format.toLowerCase()] || "image/jpeg";
 
     const message = await anthropic.messages.create({
       model: "claude-3-5-sonnet-20241022",
@@ -82,7 +57,7 @@ export async function analyzeImage(
             },
             {
               type: "text",
-              text: `You are looking at a screenshot of one or more 4chan posts. 4chan is an anonymous imageboard where users can post text and images. Posts typically have an ID number, timestamp, and optional poster name. Some text lines start with ">" which makes them appear green ("greentext"). Extract all information from this 4chan-style post(s) into JSON format. There may be multiple posts in the image. For each post, capture the post ID, board name, timestamp, and poster name. For posts with images, include the image details. Separate the content into arrays of greentext (lines starting with ">") and regular text. Lastly make a determination of whether the post contains NSFW content. Use this structure:
+              text: `You are looking at a screenshot of one or more 4chan posts. 4chan is an anonymous imageboard where users can post text and images. Posts typically have an ID number, timestamp, and optional poster name. Some text lines start with ">" which makes them appear green ("greentext"). Extract all information from this 4chan-style post(s) into JSON format. There may be multiple posts in the image. For each post, capture the post ID, board name, timestamp, and poster name. For posts with images, include the image details. Separate the content into arrays of greentext (lines starting with ">") and regular text. Lastly make a determination of whether the post contains NSFW content. DO NOT INCLUDE ANYTHING OTHER THAN THE JSON. Use this structure:
               {
                 "posts": [
                   {
