+++ backend/services/openaiService.ts
@@ -6,37 +6,43 @@ const openai = new OpenAI({
 });
 
 export async function generateEmbeddings(analysis: ImageAnalysis) {
-  const results = await Promise.all(
-    analysis.posts.map(async (post) => {
-      const fullText = [...post.content.greentext, ...post.content.text].join(
-        " "
-      );
-      const embedding = await openai.embeddings.create({
-        model: "text-embedding-3-small",
-        input: fullText,
-        encoding_format: "float",
-      });
+  try {
+    const embeddings = await Promise.all(
+      analysis.posts.map(async (post) => {
+        const fullText = [...post.content.greentext, ...post.content.text].join(
+          " "
+        );
+        const { data } = await openai.embeddings.create({
+          model: "text-embedding-3-small",
+          input: fullText,
+          encoding_format: "float",
+        });
 
-      console.log(
-        `First 5 dimensions for post ${post.post_id}:`,
-        embedding.data[0].embedding.slice(0, 5)
-      );
+        return {
+          post_id: post.post_id,
+          embedding: data[0].embedding,
+        };
+      })
+    );
 
-      return {
-        post_id: post.post_id,
-        embedding: embedding.data[0].embedding,
-      };
-    })
-  );
-  return results;
+    return embeddings;
+  } catch (error) {
+    console.error("Failed to generate embeddings:", error);
+    throw new Error("Embedding generation failed");
+  }
 }
 
 export async function generateSearchEmbedding(searchQuery: string) {
-  const embedding = await openai.embeddings.create({
-    model: "text-embedding-3-small",
-    input: searchQuery,
-    encoding_format: "float",
-  });
+  try {
+    const { data } = await openai.embeddings.create({
+      model: "text-embedding-3-small",
+      input: searchQuery,
+      encoding_format: "float",
+    });
 
-  return embedding.data[0].embedding;
+    return data[0].embedding;
+  } catch (error) {
+    console.error("Search embedding generation failed:", error);
+    throw new Error("Failed to generate search embedding");
+  }
 }
